Move LDAP components to ldap folder
